Fixes Stack Overfglow on Number conversion
I noticed a bug in the current implementation of AbstractConverter,
causing a stack overflow. It is easily reproducible;

'
SI.METRE.getConverterTo(US.FOOT).convert(1.0);
SI.METRE.getConverterTo(US.FOOT).convert(new Double(1.0));
'
Using the converter that uses promitives number works as expected.
However, using any implementation of java.lang.Number as the parameter
causes a StackOverflow. This is due to the implementation of
convert(Number) in AbstractConverter.java.

This adds a fix and enhance a unit test to prevent regression.